rag-chatbot/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
anthropic==0.7.8
sentence-transformers==2.2.2
psycopg2-binary==2.9.9
asyncpg==0.29.0
pgvector==0.2.4
pypdf2==3.0.1
python-dotenv==1.0.0
numpy==1.24.3